Refactor parser into parsers/commentv1
Moved parser logic to `parsers/commentv1` and registry to `parsers`. Updated default parser name to `commentv1`.
This structure avoids circular imports and allows for multiple parser implementations.

Changes:
- Renamed `parser` package to `parsers`.
- Created `parsers/commentv1` for the comment parser implementation.
- Updated imports and references across the codebase.
- Updated default parser name in CLI to `commentv1`.
