+++ cmd/gosubc/templates/generate_usage.txt
@@ -9,4 +9,4 @@ Subcommands:
 Flags:
     --dir string           Project root directory containing go.mod (default: .)
     --man-dir string       Directory to generate man pages in optional
-    --parser-name string   Name of the parser to use (default: comment)
+    --parser-name string   Name of the parser to use (default: commentv1)
