+++ issue55_test.go
@@ -3,7 +3,7 @@ package go_subcommand
 import (
 	"testing"
 
-	"github.com/arran4/go-subcommand/parser"
+	"github.com/arran4/go-subcommand/parsers"
 )
 
 func TestIssue55_SanitizeToIdentifier(t *testing.T) {
@@ -38,7 +38,7 @@ func TestIssue55_SanitizeToIdentifier(t *testing.T) {
 
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			if got := parser.SanitizeToIdentifier(tt.input); got != tt.expected {
+			if got := parsers.SanitizeToIdentifier(tt.input); got != tt.expected {
 				t.Errorf("SanitizeToIdentifier(%q) = %q, want %q", tt.input, got, tt.expected)
 			}
 		})
@@ -65,7 +65,7 @@ func TestIssue55_ToKebabCase(t *testing.T) {
 
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			if got := parser.ToKebabCase(tt.input); got != tt.expected {
+			if got := parsers.ToKebabCase(tt.input); got != tt.expected {
 				t.Errorf("ToKebabCase(%q) = %q, want %q", tt.input, got, tt.expected)
 			}
 		})
