+++ parsers/commentv1/parser_order_test.go
@@ -1,4 +1,4 @@
-package parser
+package commentv1
 
 import (
 	"embed"
