+++ parsers/commentv1/parser_regr_test.go
@@ -1,4 +1,4 @@
-package parser
+package commentv1
 
 import (
 	"bytes"
