+++ parsers/commentv1/testdata/parser_regr/gofmt_flags.txtar
