+++ parsers/sanitize_test.go
@@ -1,4 +1,4 @@
-package parser
+package parsers
 
 import (
 	"testing"
